+++ packages/frontend-tk-plugin-aptx/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @aptx/frontend-tk-plugin-aptx
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @aptx/frontend-tk-core@0.1.8
+
 ## 0.1.7
 
 ### Patch Changes